requirements.txt
asgiref==3.5.2
Django==4.1.3
django-phonenumber-field==7.0.0
phonenumbers==8.13.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
